Centralized filter NQL handling and page-level semantics tests
ref https://linear.app/ghost/issue/BER-3414
The remaining risk was duplicated NQL logic and weak UI-semantic coverage, so this slice moves active query conversion into one shared module and adds page tests for the members and comments behaviors the PRD calls out explicitly.
