+++ apps/posts/src/views/members/hooks/use-members-filter-state.ts
@@ -2,8 +2,8 @@ import {useMemo} from 'react';
 import {useBrowseSettings} from '@tryghost/admin-x-framework/api/settings';
 import type {Filter} from '@tryghost/shade';
 import {deriveFilterFlags} from '@src/views/filters/filter-flags';
+import {buildMemberNqlFilter, parseMemberNqlFilterParam} from '@src/views/filters/filter-nql';
 import {isMemberField, isMemberOperatorForField, MEMBER_STATIC_FIELDS} from '@src/views/filters/member-fields';
-import {buildMemberNqlFilter, parseMemberNqlFilterParam} from '@src/views/filters/member-nql';
 import {parsePredicateParams, serializePredicateParams} from '@src/views/filters/url-predicate-params';
 import {UrlFilterStateOptions, useUrlFilterState} from '@src/views/filters/use-url-filter-state';
 import {getSiteTimezone} from '@src/utils/get-site-timezone';
